enhancement of Pump sensor data
Idea was to predict machine failure. Exploratory data analysis was done inclusive data cleaning and feature engineering.
Due to a heavy imbalanced dataset all ML Models can not predict machine failure.
Idea is to build a Deep Learning LSTM Model which predicts days a head and compares it than to the actual data. If bigger variations are getting detected based on the prediction and the actual time series we would have a machine failure...